# openapi/spec.yml
openapi: 3.0.0
info:
  contact:
    name: Canva Developer Community
    url: https://community.canva.dev/
  description: API for building integrations with Canva via a REST api
  license:
    name: ©2023 All Rights Reserved
  termsOfService: https://www.canva.com/trust/legal/
  title: Canva Connect API
  version: latest
servers:
- description: Canva Connect API
  url: https://api.canva.com/rest
...
paths:
...
  /v1/asset-uploads:
    post:
      description: |-
        Starts a new job to upload an asset to a user's content library.

        The request format for this endpoint is an `application/octet-stream` body of bytes. Attach
        information about the upload using an `Asset-Upload-Metadata` header.

        You can check the status and get the results of asset upload jobs created with this API using the [Get asset upload job API](https://www.canva.dev/docs/connect/api-reference/assets/get-asset-upload-job/).
      operationId: CreateAssetUploadJob
      parameters:
      - content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetUploadMetadata'
        in: header
        name: Asset-Upload-Metadata
        required: true
      requestBody:
        content:
          application/octet-stream:
            schema:
              format: binary
              type: string
        description: Binary of the asset to upload.
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                in_progress:
                  $ref: '#/components/examples/InProgressAssetUploadJobExample'
                success:
                  $ref: '#/components/examples/SuccessAssetUploadJobExample'
                failed:
                  $ref: '#/components/examples/FailedAssetUploadJobExample'
              schema:
                $ref: '#/components/schemas/CreateAssetUploadJobResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - asset:write
      tags:
      - asset
      x-rate-limit-per-client-user: 30
  /v1/asset-uploads/{jobId}:
    get:
      description: |-
        Get the result of an asset upload job that was created using the [Create asset upload job API](https://www.canva.dev/docs/connect/api-reference/assets/create-asset-upload-job/).

        You might need to make multiple requests to this endpoint until you get a `success` or `failed` status.
      operationId: GetAssetUploadJob
      parameters:
      - description: The asset upload job ID.
        explode: false
        in: path
        name: jobId
        required: true
        schema:
          pattern: "^[a-zA-Z0-9_-]{1,50}$"
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                in_progress:
                  $ref: '#/components/examples/InProgressAssetUploadJobExample'
                success:
                  $ref: '#/components/examples/SuccessAssetUploadJobExample'
                failed:
                  $ref: '#/components/examples/FailedAssetUploadJobExample'
              schema:
                $ref: '#/components/schemas/GetAssetUploadJobResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - asset:read
      tags:
      - asset
      x-rate-limit-per-client-user: 180
  /v1/autofills:
    post:
      description: |-
        <Warning>

        Later this year, we'll be updating all brand template IDs. If your integration stores brand template IDs, you'll need to migrate to the new IDs. Once we implement this change, you'll have 6 months to update the IDs.

        </Warning>

        <Note>

        To use this API, your integration must act on behalf of a user that's a member of a [Canva Enterprise](https://www.canva.com/enterprise/) organization.

        </Note>

        Starts a new job to autofill a Canva design using a brand template and input data.

        To get a list of input data fields, use the [Get brand template dataset
        API](https://www.canva.dev/docs/connect/api-reference/brand-templates/get-brand-template-dataset/).

        Brand templates contain references to image media. You can reference image media by providing the value of `type` with `image` within the `data` object. The number of
        references to image media within a brand template cannot currently exceed 300 references.
      operationId: createDesignAutofillJob
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDesignAutofillJobRequest'
      responses:
        "200":
          content:
            application/json:
              examples:
                in_progress:
                  $ref: '#/components/examples/InProgressAutofillJobExample'
                success:
                  $ref: '#/components/examples/SuccessAutofillJobExample'
                failed:
                  $ref: '#/components/examples/FailedAutofillJobExample'
              schema:
                $ref: '#/components/schemas/CreateDesignAutofillJobResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - design:content:write
      tags:
      - autofill
      x-rate-limit-per-client-user: 10
  /v1/autofills/{jobId}:
    get:
      description: |-
        <Note>

        To use this API, your integration must act on behalf of a user that's a member of a [Canva Enterprise](https://www.canva.com/enterprise/) organization.

        </Note>

        Get the result of a design autofill job that was created using the [Create a design autofill job
        API](https://www.canva.dev/docs/connect/api-reference/autofills/create-design-autofill-job/).

        You might need to make multiple requests to this endpoint until you get a `success` or
        `failed` status.
      operationId: getDesignAutofillJob
      parameters:
      - description: The design autofill job ID.
        explode: false
        in: path
        name: jobId
        required: true
        schema:
          pattern: "^[a-zA-Z0-9_-]{1,50}$"
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                in_progress:
                  $ref: '#/components/examples/InProgressAutofillJobExample'
                success:
                  $ref: '#/components/examples/SuccessAutofillJobExample'
                failed:
                  $ref: '#/components/examples/FailedAutofillJobExample'
              schema:
                $ref: '#/components/schemas/GetDesignAutofillJobResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - design:meta:read
      tags:
      - autofill
      x-rate-limit-per-client-user: 60
  /v1/brand-templates:
    get:
      description: |-
        <Warning>

        Later this year, we'll be updating all brand template IDs. If your integration stores brand template IDs, you'll need to migrate to the new IDs. Once we implement this change, you'll have 6 months to update the IDs.

        </Warning>

        <Note>

        To use this API, your integration must act on behalf of a user that's a member of a [Canva Enterprise](https://www.canva.com/enterprise/) organization.

        </Note>

        Get a list of the [brand templates](https://www.canva.com/help/publish-team-template/) the user has access to.
      operationId: listBrandTemplates
      parameters:
      - description: Lets you search the brand templates available to the user using
          a search term or terms.
        explode: true
        in: query
        name: query
        required: false
        schema:
          type: string
        style: form
      - description: |-
          If the success response contains a continuation token, the user has access to more
          brand templates you can list. You can use this token as a query parameter and retrieve
          more templates from the list, for example
          `/v1/brand-templates?continuation={continuation}`.
          To retrieve all the brand templates available to the user, you might need to make
          multiple requests.
        explode: true
        in: query
        name: continuation
        required: false
        schema:
          type: string
        style: form
      - description: |-
          Filter the list of brand templates based on the user's ownership of the brand templates.
          This can be one of the following:

          - `any`: (Default) Brand templates owned by and shared with the user.
          - `owned`: Brand templates owned by the user.
          - `shared`: Brand templates shared with the user.
        explode: true
        in: query
        name: ownership
        required: false
        schema:
          $ref: '#/components/schemas/OwnershipType'
        style: form
      - description: |-
          Sort the list of brand templates. This can be one of the following:

          - `relevance`: (Default) Sort results using a relevance algorithm.
          - `modified_descending`: Sort results by the date last modified in descending order.
          - `modified_ascending`: Sort results by the date last modified in ascending order.
          - `title_descending`: Sort results by title in descending order.
          - `title_ascending`: Sort results by title in ascending order.
        explode: true
        in: query
        name: sort_by
        required: false
        schema:
          $ref: '#/components/schemas/SortByType'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListBrandTemplatesResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - brandtemplate:meta:read
      tags:
      - brand_template
...
  /v1/brand-templates/{brandTemplateId}/dataset:
    get:
      description: |-
        <Warning>

        Later this year, we'll be updating all brand template IDs. If your integration stores brand template IDs, you'll need to migrate to the new IDs. Once we implement this change, you'll have 6 months to update the IDs.

        </Warning>

        <Note>

        To use this API, your integration must act on behalf of a user that's a member of a [Canva Enterprise](https://www.canva.com/enterprise/) organization.

        </Note>

        Gets the dataset definition of a brand template. If the brand
        template contains autofill data fields, this API returns an object with the data field
        names and the type of data they accept.

        You can autofill a brand template using the [Create a design autofill job
        API](https://www.canva.dev/docs/connect/api-reference/autofills/create-design-autofill-job/).
      operationId: getBrandTemplateDataset
      parameters:
      - description: The brand template ID.
        explode: false
        in: path
        name: brandTemplateId
        required: true
        schema:
          pattern: "^[a-zA-Z0-9_-]{1,50}$"
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBrandTemplateDatasetResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - brandtemplate:content:read
      tags:
      - brand_template
...
  /v1/designs:
    get:
      description: |-
        Lists metadata for all the designs in a Canva user's
        [projects](https://www.canva.com/help/find-designs-and-folders/). You can also:

        - Use search terms to filter the listed designs.
        - Show designs either created by, or shared with the user.
        - Sort the results.
      operationId: listDesigns
      parameters:
      - description: "Lets you search the user's designs, and designs shared with\
          \ the user, using a search term or terms."
        example: party invites
        explode: true
        in: query
        name: query
        required: false
        schema:
          type: string
        style: form
      - description: |-
          If the success response contains a continuation token, the list contains more designs
          you can list. You can use this token as a query parameter and retrieve more
          designs from the list, for example
          `/v1/designs?continuation={continuation}`.

          To retrieve all of a user's designs, you might need to make multiple requests.
        example: RkFGMgXlsVTDbMd:MR3L0QjiaUzycIAjx0yMyuNiV0OildoiOwL0x32G4NjNu4FwtAQNxowUQNMMYN
        explode: true
        in: query
        name: continuation
        required: false
        schema:
          type: string
        style: form
      - description: |-
          Filter the list of designs based on the user's ownership of the designs.
          This can be one of the following:

          - `owned`: Designs owned by the user.
          - `shared`: Designs shared with the user.
          - `any`: Designs owned by and shared with the user.
        example: any
        explode: true
        in: query
        name: ownership
        required: false
        schema:
          $ref: '#/components/schemas/OwnershipType'
        style: form
      - description: |-
          Sort the list of designs.
          This can be one of the following:

          - `relevance`: (Default) Sort results using a relevance algorithm.
          - `modified_descending`: Sort results by the date last modified in descending order.
          - `modified_ascending`: Sort results by the date last modified in ascending order.
          - `title_descending`: Sort results by title in descending order.
          - `title_ascending`: Sort results by title in ascending order.
        example: relevance
        explode: true
        in: query
        name: sort_by
        required: false
        schema:
          $ref: '#/components/schemas/SortByType'
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetListDesignResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - design:meta:read
      tags:
      - design
    post:
      description: |-
        Creates a new Canva design. To create a new design, you can either:

        - Use a preset design type.
        - Set height and width dimensions for a custom design.

        Additionally, you can also provide the `asset_id` of an asset in the user's [projects](https://www.canva.com/help/find-designs-and-folders/) to add to the new design. To list the assets in a folder in the user's projects, use the [List folder items API](https://www.canva.dev/docs/connect/api-reference/folders/list-folder-items/).
      operationId: createDesign
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDesignRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateDesignResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - design:content:write
      tags:
      - design
      x-rate-limit-per-client-user: 20
  /v1/designs/{designId}:
    get:
      description: "Gets the metadata for a design. This includes owner information,\
        \ URLs for editing and viewing, and thumbnail information."
      operationId: getDesign
      parameters:
      - description: The design ID.
        explode: false
        in: path
        name: designId
        required: true
        schema:
          pattern: "^[a-zA-Z0-9_-]{1,50}$"
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetDesignResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error Response
      security:
      - oauthAuthCode:
        - design:meta:read
      tags:
      - design
...
components:
  examples:
    InProgressAssetUploadJobExample:
      summary: In progress job
      value:
        job:
          id: e08861ae-3b29-45db-8dc1-1fe0bf7f1cc8
          status: in_progress
    SuccessAssetUploadJobExample:
      summary: Successfully completed job
      value:
        job:
          id: e08861ae-3b29-45db-8dc1-1fe0bf7f1cc8
          status: success
          asset:
            id: Msd59349ff
            name: My Awesome Upload
            tags:
            - image
            - holiday
            - best day ever
            created_at: 1377396000
            updated_at: 1692928800
            thumbnail:
              width: 595
              height: 335
              url: https://document-export.canva.com/Vczz9/zF9vzVtdADc/2/thumbnail/0001.png?<query-string>
    FailedAssetUploadJobExample:
      summary: Failed job
      value:
        job:
          id: e08861ae-3b29-45db-8dc1-1fe0bf7f1cc8
          status: failed
          error:
            code: file_too_big
            message: Failed to import because the file is too big
    InProgressAutofillJobExample:
      summary: In progress job
      value:
        job:
          id: 450a76e7-f96f-43ae-9c37-0e1ce492ac72
          status: in_progress
    SuccessAutofillJobExample:
      summary: Successfully completed job
      value:
        job:
          id: 450a76e7-f96f-43ae-9c37-0e1ce492ac72
          status: success
          result:
            type: create_design
            design:
              id: DAFVztcvd9z
              title: My summer holiday
              url: https://www.canva.com/design/DAFVztcvd9z/edit
              thumbnail:
                width: 595
                height: 335
                url: https://document-export.canva.com/Vczz9/zF9vzVtdADc/2/thumbnail/0001.png?<query-string>
    FailedAutofillJobExample:
      summary: Failed job
      value:
        job:
          id: 450a76e7-f96f-43ae-9c37-0e1ce492ac72
          status: failed
          error:
            code: autofill_error
            message: Error autofilling design from brand template
...
  schemas:
...
    CreateAssetUploadJobResponse:
      properties:
        job:
          $ref: '#/components/schemas/AssetUploadJob'
      required:
      - job
      type: object
    GetAssetUploadJobResponse:
      properties:
        job:
          $ref: '#/components/schemas/AssetUploadJob'
      required:
      - job
      type: object
    AssetUploadJob:
      description: The status of the asset upload job.
      example:
        id: e08861ae-3b29-45db-8dc1-1fe0bf7f1cc8
        status: success
        asset:
          id: Msd59349ff
          name: My Awesome Upload
          tags:
          - image
          - holiday
          - best day ever
          created_at: 1377396000
          updated_at: 1692928800
          thumbnail:
            width: 595
            height: 335
            url: https://document-export.canva.com/Vczz9/zF9vzVtdADc/2/thumbnail/0001.png?<query-string>
      properties:
        id:
          description: The ID of the asset upload job.
          example: e08861ae-3b29-45db-8dc1-1fe0bf7f1cc8
          type: string
        status:
          $ref: '#/components/schemas/AssetUploadStatus'
        error:
          $ref: '#/components/schemas/AssetUploadError'
        asset:
          $ref: '#/components/schemas/Asset'
      required:
      - id
      - status
      type: object
    AssetUploadStatus:
      description: Status of the asset upload job.
      enum:
      - failed
      - in_progress
      - success
      example: success
      type: string
...
    AssetUploadMetadata:
      description: Metadata for the asset being uploaded.
      properties:
        name_base64:
          description: |-
            The asset's name, encoded in Base64.

            The maximum length of an asset name in Canva (unencoded) is 50 characters.

            Base64 encoding allows names containing emojis and other special
            characters to be sent using HTTP headers.
            For example, "My Awesome Upload 🚀" Base64 encoded
            is `TXkgQXdlc29tZSBVcGxvYWQg8J+agA==`.
          example: TXkgQXdlc29tZSBVcGxvYWQg8J+agA==
          minLength: 1
          type: string
      required:
      - name_base64
      type: object
    CreateDesignAutofillJobRequest:
      properties:
        brand_template_id:
          description: ID of the input brand template.
          example: DAFVztcvd9z
          type: string
        title:
          description: |-
            Title to use for the autofilled design. Must be less than 256 characters.

            If no design title is provided, the autofilled design will have the same title as the brand template.
          maxLength: 255
          minLength: 1
          type: string
        data:
          additionalProperties:
            $ref: '#/components/schemas/DatasetValue'
          description: Data object containing the data fields and values to autofill.
          example:
            cute_pet_image_of_the_day:
              type: image
              asset_id: Msd59349ff
            cute_pet_witty_pet_says:
              type: text
              text: It was like this when I got here!
          type: object
      required:
      - brand_template_id
      - data
      type: object
    CreateDesignAutofillJobResponse:
      properties:
        job:
          $ref: '#/components/schemas/DesignAutofillJob'
      required:
      - job
      type: object
    DatasetValue:
      description: The data field to autofill.
      discriminator:
        mapping:
          image: '#/components/schemas/DatasetImageValue'
          text: '#/components/schemas/DatasetTextValue'
        propertyName: type
      oneOf:
      - $ref: '#/components/schemas/DatasetImageValue'
      - $ref: '#/components/schemas/DatasetTextValue'
      type: object
...
    GetDesignAutofillJobResponse:
      properties:
        job:
          $ref: '#/components/schemas/DesignAutofillJob'
      required:
      - job
      type: object
    DesignAutofillJob:
      description: Details about the autofill job.
      properties:
        id:
          description: ID of the asynchronous job that is creating the design using
            the provided data.
          example: 450a76e7-f96f-43ae-9c37-0e1ce492ac72
          type: string
        status:
          $ref: '#/components/schemas/DesignAutofillStatus'
        result:
          $ref: '#/components/schemas/DesignAutofillJobResult'
        error:
          $ref: '#/components/schemas/AutofillError'
      required:
      - id
      - status
      type: object
...
    ListBrandTemplatesResponse:
      properties:
        continuation:
          description: |-
            If the success response contains a continuation token, the user has access to more
            brand templates you can list. You can use this token as a query parameter and retrieve
            more templates from the list, for example
            `/v1/brand-templates?continuation={continuation}`.
            To retrieve all the brand templates available to the user, you might need to make
            multiple requests.
          example: RkFGMgXlsVTDbMd:MR3L0QjiaUzycIAjx0yMyuNiV0OildoiOwL0x32G4NjNu4FwtAQNxowUQNMMYN
          type: string
        items:
          description: The list of brand templates.
          items:
            $ref: '#/components/schemas/BrandTemplate'
          type: array
      required:
      - items
      type: object
...
    GetBrandTemplateDatasetResponse:
      description: Successful response from a `getBrandTemplateDataset` request.
      properties:
        dataset:
          additionalProperties:
            $ref: '#/components/schemas/DataField'
          description: |-
            The dataset definition for the brand template. The dataset definition contains the data inputs available for use with the
            [Create design autofill job API](https://www.canva.dev/docs/connect/api-reference/autofills/create-design-autofill-job/).
          example:
            cute_pet_image_of_the_day:
              type: image
            cute_pet_witty_pet_says:
              type: text
          type: object
      type: object
...
    SortByType:
      enum:
      - relevance
      - modified_descending
      - modified_ascending
      - title_descending
      - title_ascending
      type: string
    OwnershipType:
      enum:
      - any
      - owned
      - shared
      type: string
    GetListDesignResponse:
      properties:
        continuation:
          description: |-
            A continuation token.
            If the success response contains a continuation token, the list contains more designs
            you can list. You can use this token as a query parameter and retrieve more
            designs from the list, for example
            `/v1/designs?continuation={continuation}`.

            To retrieve all of a user's designs, you might need to make multiple requests.
          example: RkFGMgXlsVTDbMd:MR3L0QjiaUzycIAjx0yMyuNiV0OildoiOwL0x32G4NjNu4FwtAQNxowUQNMMYN
          type: string
        items:
          description: The list of designs.
          items:
            $ref: '#/components/schemas/Design'
          type: array
      required:
      - items
      type: object
    CreateDesignRequest:
      description: |-
        Body parameters for creating a new design.
        At least one of `design_type` or `asset_id` must be defined
        to create a new design.
      properties:
        design_type:
          $ref: '#/components/schemas/DesignTypeInput'
        asset_id:
          description: The ID of an asset to insert into the created design.
          example: Msd59349ff
          type: string
        title:
          description: The name of the design. Must be less than 256 characters.
          example: My Holiday Presentation
          maxLength: 255
          minLength: 1
          type: string
      type: object
    CreateDesignResponse:
      description: Details about the new design.
      properties:
        design:
          $ref: '#/components/schemas/Design'
      required:
      - design
      type: object
    GetDesignResponse:
      description: Successful response from a `getDesign` request.
      properties:
        design:
          $ref: '#/components/schemas/Design'
      required:
      - design
      type: object
...
    Error:
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          description: A human-readable description of what went wrong.
          type: string
      required:
      - code
      - message
      type: object
...
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http
    oauthAuthCode:
      flows:
        authorizationCode:
          authorizationUrl: https://www.canva.com/api/oauth/authorize
          scopes:
            design:content:read: View the contents of the user's designs.
            design:meta:read: View the metadata of the user's designs.
            design:content:write: Create designs on the user's behalf.
            folder:read: |-
              View the metadata and contents of the user's folders, including their **Projects**
              folder.
            folder:write: |-
              Add, move, or remove the user's folders. It also lets you edit folder metadata,
              such as the folder's name.
            asset:read: "View the metadata for the user's assets, such as uploaded\
              \ images."
            asset:write: "Upload, update, or delete assets on the user's behalf."
            comment:read: "View the comments on the user's designs, and the associated\
              \ metadata."
            comment:write: Create comments and replies on the user's designs.
            collaboration:event: Receive webhook notifications about events relevant
              to the user.
            brandtemplate:meta:read: View the metadata of the brand templates associated
              with the user's brand.
            brandtemplate:content:read: Read the content of the brand templates associated
              with the user's brand.
            profile:read: Read a user's profile and account information.
          tokenUrl: https://api.canva.com/rest/v1/oauth/token
      type: oauth2